+++ include/fb/game/stat.h
@@ -35,10 +35,10 @@ class stat
 {
 private:
     life&    owner;
-    uint32_t _hp                = 0;
-    uint32_t _mp                = 0;
-    int32_t  _buff_hp           = 0;
-    int32_t  _buff_mp           = 0;
+    uint64_t _hp                = 0;
+    uint64_t _mp                = 0;
+    int64_t  _buff_hp           = 0;
+    int64_t  _buff_mp           = 0;
     uint8_t  _buff_str          = 0;
     uint8_t  _buff_dex          = 0;
     uint8_t  _buff_int          = 0;
@@ -46,7 +46,7 @@ private:
     int8_t   _buff_magdef       = 0;
     int8_t   _buff_dam          = 0;
     int8_t   _buff_hit          = 0;
-    uint32_t _buff_regenerative = 0;
+    uint64_t _buff_regenerative = 0;
 
 protected:
     stat(life& owner);
@@ -63,8 +63,8 @@ public:
 
 public:
     // clang-format off
-    virtual uint32_t base_hp() const = 0;
-    virtual uint32_t base_mp() const = 0;
+    virtual uint64_t base_hp() const = 0;
+    virtual uint64_t base_mp() const = 0;
     virtual uint8_t  base_str() const = 0;
     virtual uint8_t  base_dex() const = 0;
     virtual uint8_t  base_int() const = 0;
@@ -72,11 +72,11 @@ public:
     virtual int8_t   base_magdef() const = 0;
     virtual uint8_t  base_dam() const = 0;
     virtual uint8_t  base_hit() const = 0;
-    virtual uint32_t base_regenerative() const = 0;
-    virtual int32_t  buff_hp() const;
-    virtual void     buff_hp(int32_t value);
-    virtual int32_t  buff_mp() const;
-    virtual void     buff_mp(int32_t value);
+    virtual uint64_t base_regenerative() const = 0;
+    virtual int64_t  buff_hp() const;
+    virtual void     buff_hp(int64_t value);
+    virtual int64_t  buff_mp() const;
+    virtual void     buff_mp(int64_t value);
     virtual uint8_t  buff_str() const;
     virtual void     buff_str(uint8_t value);
     virtual uint8_t  buff_dex() const;
@@ -91,18 +91,18 @@ public:
     virtual void     buff_dam(int8_t value);
     virtual int8_t   buff_hit() const;
     virtual void     buff_hit(int8_t value);
-    virtual uint32_t buff_regenerative() const;
-    virtual void     buff_regenerative(uint32_t value);
-    virtual uint32_t hp() const;
-    virtual void     hp(uint32_t value, bool notify = true);
-    virtual uint32_t heal(uint32_t value, fb::game::object* from = nullptr, bool notify = true);
-    virtual uint32_t damage(uint32_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true);
-    virtual uint32_t mp() const;
-    virtual void     mp(uint32_t value, bool notify = true);
-    virtual uint32_t mp_up(uint32_t value, fb::game::object* from = nullptr, bool notify = true);
-    virtual uint32_t mp_down(uint32_t value, fb::game::object* from = nullptr, bool notify = true);
-    virtual uint32_t maxhp() const;
-    virtual uint32_t maxmp() const;
+    virtual uint64_t buff_regenerative() const;
+    virtual void     buff_regenerative(uint64_t value);
+    virtual uint64_t hp() const;
+    virtual void     hp(uint64_t value, bool notify = true);
+    virtual uint64_t heal(uint64_t value, fb::game::object* from = nullptr, bool notify = true);
+    virtual uint64_t damage(uint64_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true);
+    virtual uint64_t mp() const;
+    virtual void     mp(uint64_t value, bool notify = true);
+    virtual uint64_t mp_up(uint64_t value, fb::game::object* from = nullptr, bool notify = true);
+    virtual uint64_t mp_down(uint64_t value, fb::game::object* from = nullptr, bool notify = true);
+    virtual uint64_t maxhp() const;
+    virtual uint64_t maxmp() const;
     virtual uint8_t  str() const;
     virtual uint8_t  dex() const;
     virtual uint8_t  intelligence() const;
@@ -110,15 +110,15 @@ public:
     virtual int8_t   magdef() const;
     virtual int8_t   dam() const;
     virtual int8_t   hit() const;
-    virtual uint32_t regenerative() const;
+    virtual uint64_t regenerative() const;
     // clang-format on
 };
 
 class character_stat : public stat
 {
 private:
-    uint32_t _max_hp       = 0;
-    uint32_t _max_mp       = 0;
+    uint64_t _max_hp       = 0;
+    uint64_t _max_mp       = 0;
     uint8_t  _str          = 0;
     uint8_t  _dex          = 0;
     uint8_t  _int          = 0;
@@ -126,7 +126,7 @@ private:
     int32_t  _magdef       = 0;
     int32_t  _dam          = 0;
     int32_t  _hit          = 0;
-    uint32_t _regenerative = 0;
+    uint64_t _regenerative = 0;
 
 public:
     character& owner;
@@ -137,8 +137,8 @@ public:
 
 public:
     // clang-format off
-    void             base_hp(uint32_t value, bool notify = true);
-    void             base_mp(uint32_t value, bool notify = true);
+    void             base_hp(uint64_t value, bool notify = true);
+    void             base_mp(uint64_t value, bool notify = true);
     void             base_str(uint8_t value, bool notify = true);
     void             base_dex(uint8_t value, bool notify = true);
     void             base_int(uint8_t value, bool notify = true);
@@ -146,9 +146,9 @@ public:
     void             base_magdef(int8_t value, bool notify = true);
     void             base_dam(uint8_t value, bool notify = true);
     void             base_hit(uint8_t value, bool notify = true);
-    void             base_regenerative(uint32_t value, bool notify = true);
-    uint32_t         base_hp() const;
-    uint32_t         base_mp() const;
+    void             base_regenerative(uint64_t value, bool notify = true);
+    uint64_t         base_hp() const;
+    uint64_t         base_mp() const;
     uint8_t          base_str() const;
     virtual uint8_t  str() const;
     uint8_t          base_dex() const;
@@ -163,11 +163,11 @@ public:
     virtual int8_t   dam() const;
     uint8_t          base_hit() const;
     virtual int8_t   hit() const;
-    uint32_t         base_regenerative() const;
-    virtual uint32_t maxhp() const override;
-    virtual uint32_t maxmp() const override;
-    virtual uint32_t regenerative() const override;
-    uint32_t         damage(uint32_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true) override final;
+    uint64_t         base_regenerative() const;
+    virtual uint64_t maxhp() const override;
+    virtual uint64_t maxmp() const override;
+    virtual uint64_t regenerative() const override;
+    uint64_t         damage(uint64_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true) override final;
     // clang-format on
 };
 
@@ -186,8 +186,8 @@ public:
 
 public:
     // clang-format off
-    uint32_t base_hp() const override final;
-    uint32_t base_mp() const override final;
+    uint64_t base_hp() const override final;
+    uint64_t base_mp() const override final;
     uint8_t  base_str() const override final;
     uint8_t  base_dex() const override final;
     uint8_t  base_int() const override final;
@@ -195,11 +195,11 @@ public:
     int8_t   base_magdef() const override final;
     uint8_t  base_dam() const override final;
     uint8_t  base_hit() const override final;
-    uint32_t base_regenerative() const override final;
-    uint32_t hp() const override final;
-    uint32_t damage(uint32_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true) override final;
-    void     hp(uint32_t value, bool notify = true) override final;
-    uint32_t heal(uint32_t value, fb::game::object* from = nullptr, bool notify = true) override final;
+    uint64_t base_regenerative() const override final;
+    uint64_t hp() const override final;
+    uint64_t damage(uint64_t value, std::shared_ptr<fb::game::object> from = nullptr, bool critical = false, float rate = 1.0f, bool physical = true, bool fixed = false, bool notify = true) override final;
+    void     hp(uint64_t value, bool notify = true) override final;
+    uint64_t heal(uint64_t value, fb::game::object* from = nullptr, bool notify = true) override final;
     // clang-format on
 };
